ãadd entitiy

+++ src/main/java/com/example/SpringJWT/jwt/JWTFilter.java
@@ -42,7 +42,7 @@ public class JWTFilter extends OncePerRequestFilter {
             return;
         }
 
-        Integer id = jwtUtil.getId(token);
+        long id = jwtUtil.getId(token);
         String role = jwtUtil.getRole(token);
 
         JWTTmp jwtTmp = new JWTTmp(id, role);
